requirements.txt
requests
pytest
structlog
curlify
pydantic
pyhamcrest
